+++ v1_21/src/main/java/net/countercraft/movecraft/support/v1_21/IAsyncChunk.java
@@ -1,4 +1,4 @@
-package net.countercraft.movecraft.support.v1_21_1;
+package net.countercraft.movecraft.support.v1_21;
 
 import com.google.common.cache.CacheBuilder;
 import com.google.common.cache.CacheLoader;
